r11-01=12:10
first push. Contains a work in progress of the fighter profile. I got
the text views to read from the .java correctly. Currently using an
array list with dummy values until we sync it with the backend. Now I
have to flesh out the front end and add picture support